feat: first update ssr pages
