+++ class-components/src/__tests__/App.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import App from './App'
+import App from '../pages/_app'
 
 describe('App component', () => {
   it('renders SearchPage for the root route', () => {
