+++ class-components/src/app/[[...slug]]/client.tsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import dynamic from 'next/dynamic'
 
-const App = dynamic(() => import('../../App'), { ssr: false })
+const App = dynamic(() => import('../../pages/_app'), { ssr: false })
 
 export function ClientOnly() {
   return <App />
