+++ class-components/src/context/ThemeProvider.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactNode, useState } from 'react'
 import { createContext } from 'react'
