+++ class-components/src/pages/details/detailsPage.scss
